Relocated View, ViewModel, UiState, and UiEvent from teamsScreen to createEditScreen.
