+++ app/src/main/java/com/oracle/visualize/presentation/screens/mainScreen/MainView.kt
@@ -15,9 +15,9 @@ import androidx.navigation.compose.rememberNavController
 import com.oracle.visualize.domain.models.NavRoutes
 import com.oracle.visualize.presentation.components.BottomNavBar
 import com.oracle.visualize.presentation.screens.createChartScreen.CreatePage
+import com.oracle.visualize.presentation.screens.createEditScreen.CreateEditTeamPage
 import com.oracle.visualize.presentation.screens.feedScreen.FeedPage
 import com.oracle.visualize.presentation.screens.notificationScreen.NotificationPage
-import com.oracle.visualize.presentation.screens.teamsScreen.CreateEditTeamPage
 import com.oracle.visualize.presentation.screens.teamsScreen.TeamsPage
